backend/requirements.txt
flask
flask-cors
mongoengine
pymongo[srv]
PyJWT==1.7.1